v3: publish recovered contracts across corpus

+++ v3/silo-finance/src/SiloHookV1-2/silo-core/contracts/incentives/lib/DistributionTypes.sol
@@ -1,17 +1,16 @@
 // SPDX-License-Identifier: agpl-3.0
-
 pragma solidity 0.8.28;
 
 library DistributionTypes {
     struct IncentivesProgramCreationInput {
         string name;
         address rewardToken;
-        uint256 emissionPerSecond;
+        uint104 emissionPerSecond;
         uint40 distributionEnd;
     }
 
     struct AssetConfigInput {
-        uint256 emissionPerSecond;
+        uint104 emissionPerSecond;
         uint256 totalStaked;
         address underlyingAsset;
     }
